refactor: simplify native error bridging
